query: Add support for orderBy
Added support for sorting by introducing new query parameter 'orderBy'.
It can contain a string or an array of strings in the form of 'name' or
'-status' to sort the returned collection in ascending or descending order.

Example:
{
  select: ['id', 'name'],
  orderBy: ['status', '-name']
}

This will return the collection sorted by status (in ascending order) and
additionaly sorted by descending name, when the status is the same.